requirement.txt
numpy
decord
xformers
easydict
OpenEXR==3.3.1
click           # ==8.1.7
opencv-python   # ==4.10.0.84
trimesh         # ==4.5.1
pillow          # ==10.4.0
huggingface_hub # ==0.25.2 
torch==2.1.0
torchvision==0.16.0  # Update to match PyTorch 2.1.0
matplotlib==3.7.1
tqdm==4.65.0
imageio==2.31.1
imageio-ffmpeg==0.4.8
einops==0.6.1
timm==0.6.13
flow_vis 
tensorboard

# INSTALL monst3r
# minimal requirements for inference
roma
gdown
pyglet<2

# for camera trajectory
evo


# vggt
safetensors
gradio==5.17.1
viser==0.2.23
hydra-core
omegaconf
opencv-python
scipy
onnxruntime
requests
trimesh
matplotlib
